+++ NEWS.md
@@ -1,7 +1,18 @@
+# geneticae 0.2.0
+
+Changed default option to "symmetrical" in the SVP argument of the GGEmodel()
+function.
+
+Changed agridat package from Imports to Suggests in DESCRIPTION file.
+
 # geneticae 0.1.0
 
-Fist submission to CRAN.
+First submission to CRAN.
 
 # geneticae 0.0.9000
 
 This is the first development version of the package.
+
+
+
+